修改像素里程为double，因为配置文件中保存float乱码

+++ src/AppConfig.cpp
@@ -13,7 +13,7 @@ QString AppConfig::CameraPath_6 = "G:";
 QString AppConfig::SavePicturePath = "D:Vedio";
 QString AppConfig::SaveResultPath = "D:Vedio";
 double AppConfig::Mileage = 0.0;
-float AppConfig::Y_ACC = 0.21f;
+double AppConfig::Y_ACC = 0.21f;
 int AppConfig::MileageDown = 1;
 bool AppConfig::WriteFlag = false;
 bool AppConfig::Style = false;
@@ -35,7 +35,7 @@ void AppConfig::readConfig()
     AppConfig::SavePicturePath = set.value("SavePicturePath").toString();
     AppConfig::SaveResultPath = set.value("SaveResultPath").toString();
     AppConfig::Mileage = set.value("Mileage").toDouble();
-    AppConfig::Y_ACC = set.value("Y_ACC").toFloat();
+    AppConfig::Y_ACC = set.value("Y_ACC").toDouble();
     //AppConfig::MileageDown = set.value("Mileage").toInt();
     set.endGroup();
